refactor(painters/dom): extract image rendering helpers into modules
Move image fragment and drawing image creation out of renderer.ts into
dedicated modules under images/. Renderer now delegates:

- renderImageFragment → images/image-fragment.ts (with the
  buildImageGeometryTransform / applyImageGeometryTransform helpers it
  owned)
- createDrawingImageElement, createShapeGroupImageElement, and the
  inline shape-text image element creation → images/drawing-image.ts

No behavior change — pure extraction. Trims ~157 lines from renderer.ts.
